fix: address PR review — move DsfrPictogram, fix grid layout, update tiles
- Move DsfrPictogram from modules/home to modules/layout/shared (fix
  inverted dependency)
- Use fr-col-md-9 + fr-col-md-3 grid for equal-height tiles
- Replace "Textes de référence" tile with "Aides et ressources"
- Remove duplicate HomeResources from home page
- Add DsfrPictogram to global test mock
