+++ packages/app/src/modules/home/index.ts
@@ -1,4 +1,3 @@
-export { DsfrPictogram } from "./DsfrPictogram";
 export { HomeHero } from "./HomeHero";
 export { HomeNotice } from "./HomeNotice";
 export { HomePage } from "./HomePage";
