+++ packages/app/src/modules/layout/ResourceBanner.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import { DsfrPictogram } from "~/modules/home";
+import { DsfrPictogram } from "./shared/DsfrPictogram";
 
 type ResourceTileProps = {
 	href: string;
@@ -43,34 +43,38 @@ export function ResourceBanner() {
 			className="fr-background-alt--blue-france fr-py-7w"
 		>
 			<div className="fr-container">
-				<div className="fr-grid-row fr-grid-row--gutters">
-					<div className="fr-col-12 fr-col-md">
-						<ResourceTile
-							detail="Réponses aux questions les plus courantes"
-							href="/faq"
-							pictogramPath="/dsfr/artwork/pictograms/system/information.svg"
-							title="Questions fréquentes (FAQ)"
-						/>
-					</div>
-					<div className="fr-col-12 fr-col-md">
-						<ResourceTile
-							detail="Consultez les textes législatifs et réglementaires"
-							href="/textes-reference"
-							pictogramPath="/dsfr/artwork/pictograms/document/document.svg"
-							title="Textes de référence"
-						/>
-					</div>
-					<div className="fr-col-12 fr-col-md">
-						<ResourceTile
-							detail="Besoin d'aide ? Contactez nos services d'assistance"
-							href="/aide/nous-contacter"
-							pictogramPath="/dsfr/artwork/pictograms/digital/avatar.svg"
-							title="Nous contacter"
-						/>
+				<div className="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
+					<div className="fr-col-12 fr-col-md-9">
+						<div className="fr-grid-row fr-grid-row--gutters">
+							<div className="fr-col-12 fr-col-md-4">
+								<ResourceTile
+									detail="Réponses aux questions les plus courantes"
+									href="/faq"
+									pictogramPath="/dsfr/artwork/pictograms/system/information.svg"
+									title="Questions fréquentes (FAQ)"
+								/>
+							</div>
+							<div className="fr-col-12 fr-col-md-4">
+								<ResourceTile
+									detail="Guides, ressources et textes de référence"
+									href="/aide"
+									pictogramPath="/dsfr/artwork/pictograms/document/document.svg"
+									title="Aides et ressources"
+								/>
+							</div>
+							<div className="fr-col-12 fr-col-md-4">
+								<ResourceTile
+									detail="Besoin d'aide ? Contactez nos services d'assistance"
+									href="/aide/nous-contacter"
+									pictogramPath="/dsfr/artwork/pictograms/digital/avatar.svg"
+									title="Nous contacter"
+								/>
+							</div>
+						</div>
 					</div>
 					<div
 						aria-hidden="true"
-						className="fr-col-md-2 fr-hidden fr-unhidden-md"
+						className="fr-col-12 fr-col-md-3 fr-hidden fr-unhidden-md"
 					>
 						<Image
 							alt=""
