+++ packages/app/src/modules/layout/index.ts
@@ -2,5 +2,6 @@ export { Breadcrumb } from "./Breadcrumb";
 export { Footer } from "./Footer";
 export { Header } from "./Header";
 export { ResourceBanner } from "./ResourceBanner";
+export { DsfrPictogram } from "./shared/DsfrPictogram";
 export { NewTabNotice } from "./shared/NewTabNotice";
 export { SkipLinks } from "./shared/SkipLinks";
