+++ packages/app/src/modules/layout/shared/DsfrPictogram.tsx
